simplemps/requirements.txt
psycopg2-binary==2.8.2
attrs==19.1.0
chardet==3.0.4
configobj==5.0.6
django==2.2.1
django-environ==0.4.5
django-statici18n==1.8.3
djangorestframework==3.10.0
django-jsignature==0.9
django-extensions==3.0.9
dnspython==1.16.0
enum34==1.1.6
idna==2.8
ipaddress==1.0.22
ndg-httpsclient==0.5.1
pbr==5.2.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycrypto==2.6.1
pylint==2.3.1
pylint-django==2.0.9
pyopenssl==19.0.0
pyserial==3.4
pytz==2019.1
requests==2.22.0
service-identity==18.1.0
six==1.12.0
stevedore==1.30.1
twisted==19.2.0
urllib3==1.25.3
uwsgi==2.0.18
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
workdays==1.4
zope-interface==4.6.0
openpyxl==2.6.2
autopep8==1.4.4
django-cors-headers==3.0.1
fiscalyear==0.1.0
zeep==4.0.0
ramda==0.6.0